backend/requirements.txt
# FastAPI
fastapi[standard]==0.115.0
uvicorn[standard]==0.30.0
python-multipart==0.0.9

# Database
sqlalchemy[asyncio]==2.0.35
asyncpg==0.29.0
alembic==1.13.0

# Auth
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Redis & Celery
redis==5.0.0
celery[redis]==5.4.0

# Settings
pydantic-settings==2.5.0

# PDF & QR
reportlab==4.2.0
qrcode[pil]==7.4.0

# HTTP
httpx==0.27.0

# Testing
pytest==8.3.0
pytest-asyncio==0.24.0
